refactor: :recycle: rebuild structure of screens #3
