+++ lib/screens/home_screen.dart
@@ -1,17 +1,16 @@
 import 'package:flutter/material.dart';
-import 'package:ptt_companion/ptt_screen.dart';
+import 'package:ptt_companion/screens/ptt_screen.dart';
 import 'package:qr_flutter/qr_flutter.dart';
-import 'package:wear/wear.dart';
 
-class WatchScreen extends StatefulWidget {
-  const WatchScreen({super.key});
+class HomeScreen extends StatefulWidget {
+  const HomeScreen({Key? key}) : super(key: key);
 
   @override
-  State<WatchScreen> createState() => _WatchScreenState();
+  State<HomeScreen> createState() => _HomeScreenState();
 }
 
-class _WatchScreenState extends State<WatchScreen> {
-  String _randomCode = '';
+class _HomeScreenState extends State<HomeScreen> {
+  late String randomCode;
 
   @override
   void initState() {
@@ -21,28 +20,10 @@ class _WatchScreenState extends State<WatchScreen> {
 
   void _generateRandomCode() {
     setState(() {
-      _randomCode = 'CODE_${DateTime.now().microsecondsSinceEpoch}';
+      randomCode = 'CODE_${DateTime.now().microsecondsSinceEpoch}';
     });
   }
 
-  @override
-  Widget build(BuildContext context) => WatchShape(
-        builder: (context, shape, child) => Scaffold(
-          body: HomeScreen(_randomCode),
-        ),
-      );
-}
-
-class HomeScreen extends StatefulWidget {
-  final String randomCode;
-
-  const HomeScreen(this.randomCode, {Key? key}) : super(key: key);
-
-  @override
-  State<HomeScreen> createState() => _HomeScreenState();
-}
-
-class _HomeScreenState extends State<HomeScreen> {
   @override
   Widget build(BuildContext context) => Scaffold(
         floatingActionButton: FloatingActionButton(
@@ -66,15 +47,11 @@ class _HomeScreenState extends State<HomeScreen> {
               children: [
                 GestureDetector(
                   onTap: () => setState(() {
-                    final watchScreenState =
-                        context.findAncestorStateOfType<_WatchScreenState>();
-                    if (watchScreenState != null) {
-                      watchScreenState._generateRandomCode();
-                    }
+                    _generateRandomCode();
                   }),
                   child: QrImage(
                     foregroundColor: Colors.white,
-                    data: widget.randomCode,
+                    data: randomCode,
                     version: QrVersions.auto,
                     size: 100,
                   ),
